Split positionCmp into its own file.
